requirements.txt
absl-py==2.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
efficientnet_pytorch==0.7.1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
grpcio==1.68.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
lightning-utilities==0.11.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
munch==4.0.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pretrainedmodels==0.7.4
protobuf==5.28.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
segmentation-models-pytorch==0.3.4
setuptools==75.6.0
six==1.16.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
timm==0.9.7
torch==2.5.1
torch-tb-profiler==0.4.3
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
